Add `sig_figs` field to `StanModel`
This change enables the user to control the number of significant digits
which are preserved in the output. `sig_figs=6` is the default
cmdstan option, which is what `StanSample` has been defaulting to.

Typically, a user would should prefer to generate outputs with
`sig_figs=18` so that the f64's are uniquely identified. It might be
wise to make such a recommendation in the documentation, but I suppose
that casual users would complain about the correspondingly increased
.csv sizes (and subsequent read times).